LiverTox-RAG-LLM/requirements.txt
# Core dependencies
lxml>=4.9.0
beautifulsoup4>=4.12.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# NLP and Embeddings
sentence-transformers==2.2.2
transformers==4.33.3
torch>=2.0.0

# Vector Search
faiss-cpu>=1.7.4

# LLM Support
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0
llama-cpp-python==0.3.9

# Utilities
huggingface_hub==0.16.4
accelerate==0.21.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# Optional - for web interface
streamlit>=1.28.0